+++ astropy/visualization/__init__.py
@@ -4,6 +4,7 @@ from .hist import *
 from .interval import *
 from .mpl_normalize import *
 from .mpl_style import *
+from .scatter_contour import scatter_contour
 from .stretch import *
 from .transform import *
 from .units import *